Removed Shared Directory
Based on C835772's suggestion, pulled the files up one level so I don't
have the same shared directory name in the repo as what is built in the
vagrantfile by default. All files & dirs are now under Scenario1 and can
all be copied into the local shared directory created on the host.